requirements.txt
# ───────────────────────────── requirements.txt ─────────────────────────────
# Core LLM client
openai>=1.14,<2.0          # modern SDK with httpx built‑in

# NLP sentence splitter
spacy>=3.7,<4.0            # small‑model compatible

# Progress‑bar utility (used in preprocess_csv.py)
tqdm>=4.66,<5.0

# Read .env values for your OpenAI key (llm_utils relies on it)
python-dotenv>=1.0,<2.0
# ─────────────────────────────────────────────────────────────────────────────
